+++ .gitignore
@@ -2,3 +2,6 @@ node_modules/
 /test-results/
 /playwright-report/
 /playwright/.cache/
+/test-results/
+/playwright-report/
+/playwright/.cache/
